refactor(react): use updated hooks and utilities in all components
