+++ packages/react/src/components/Disclosure/index.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import classNames from 'classnames';
 import debounce from 'lodash.debounce';
 import isEqual from 'react-fast-compare';
-import { hasTransition, prefix } from '../../utilities';
+import { hasTransition } from '../../utilities';
+import { prefix } from '../../config';
 import { BaseDetails, BaseDetailsProps } from '../BaseDetails';
 import { BaseSummary, BaseSummaryProps } from '../BaseSummary';
 import { Icon } from '../Icon';
